# leva-backend/docs/api-specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Leva API",
    "version": "1.1.0",
    "description": "API kontrak lengkap untuk Proyek Leva — platform rekomendasi AI tools berbasis RAG dan Qdrant.\nReferensi: Dokumen Konsep Dasar dan SRS Proyek Leva (IEEE 830-1998).\n",
    "contact": {}
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/tasks": {
      "post": {
        "tags": [
          "D. Dekomposisi Tugas"
        ],
        "summary": "Submit Tugas untuk Dekomposisi Async",
        "description": "**REQ-F03** — Mengirimkan tugas (teks atau PDF) untuk diproses secara async.\nMengembalikan **HTTP 202** karena proses berjalan di background via Laravel Queue + Redis.\n\n**Job pipeline:**\n1. Ekstrak teks PDF (`spatie/pdf-to-text`) → chunk\n2. LLM dengan paradigma KERNEL prompt → parse JSON output\n3. Query Qdrant per sub-task\n4. Simpan ke `tasks_master` + `atomic_sub_tasks`\n\nPDF disimpan sementara di tmpfs dan dihapus otomatis setelah ekstraksi (REQ-NF03).\n`source_pdf_hash` (SHA-256) disimpan untuk deduplication.\n",
        "operationId": "tasks_submit",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "pdf_file": {
                    "type": "string",
                    "format": "binary",
                    "description": "File PDF (maks 10MB)"
                  },
                  "description": {
                    "type": "string",
                    "description": "Deskripsi opsional"
                  }
                }
              }
            },
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Isi tugas dalam teks bebas"
                  },
                  "description": {
                    "type": "string",
                    "description": "Deskripsi opsional"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Task diterima dan sedang diproses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Task submitted for processing"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "task_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string",
                          "example": "processing"
                        },
                        "estimated_seconds": {
                          "type": "integer",
                          "example": 25
                        },
                        "poll_url": {
                          "type": "string",
                          "example": "/api/tasks/uuid/status"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "422": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "D. Dekomposisi Tugas"
        ],
        "summary": "Daftar Semua Task Milik User",
        "operationId": "tasks_list",
        "responses": {
          "200": {
            "description": "Daftar task berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Tasks retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "tasks": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/TaskSummary"
                          }
                        },
                        "pagination": {
                          "$ref": "#/components/schemas/Pagination"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
    "/tasks/{taskId}/status": {
      "get": {
        "tags": [
          "D. Dekomposisi Tugas"
        ],
        "summary": "Polling Status Pemrosesan Task Async",
        "description": "**REQ-F03** — Fallback polling jika WebSocket belum aktif.\nFrontend melakukan polling setiap **2 detik**.\n\nAlternatif: WebSocket via Laravel Broadcasting — event `TaskCompleted`,\nchannel `private-user.{userId}`.\n",
        "operationId": "tasks_poll_status",
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Status task berhasil diambil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Task status retrieved"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "task_id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "$ref": "#/components/schemas/TaskStatus"
                        },
                        "progress_message": {
                          "type": "string",
                          "example": "Sub-tasks berhasil dibuat"
                        },
                        "sub_tasks_count": {
                          "type": "integer",
                          "example": 5
                        },
                        "ready": {
                          "type": "boolean",
                          "example": true
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Laravel Sanctum Bearer Token. Diperoleh dari POST /auth/login."
      },
      "ScraperSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Scraper-Secret",
        "description": "Secret key internal untuk webhook scraper (dari .env SCRAPER_SECRET_KEY)."
      }
    },
    "schemas": {
...
      "TaskStatus": {
        "type": "string",
        "enum": [
          "processing",
          "completed",
          "failed"
        ],
        "description": "- `processing`: Job sedang berjalan di queue\n- `completed`: Dekomposisi selesai\n- `failed`: Gagal — sertakan error_message\n"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "text",
          "pdf"
        ]
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "integer",
            "example": 1
          },
          "per_page": {
            "type": "integer",
            "example": 12
          },
          "total": {
            "type": "integer",
            "example": 48
          },
          "last_page": {
            "type": "integer",
            "example": 4
          }
        }
      },
      "ErrorMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ValidationErrorBody": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "The given data was invalid."
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": [
                "The email has already been taken."
              ]
            }
          }
        }
      },
...
      "TaskSummary": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "example": "Menyusun Skripsi Teknik Informatika"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus"
          },
          "sub_tasks_count": {
            "type": "integer",
            "example": 5
          },
          "completed_count": {
            "type": "integer",
            "example": 2
          },
          "source_type": {
            "$ref": "#/components/schemas/SourceType"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "Token tidak valid atau tidak disertakan",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorMessage"
            },
            "example": {
              "message": "Unauthenticated."
            }
          }
        }
      },
...
      "ValidationError": {
        "description": "Validasi input gagal (HTTP 422)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationErrorBody"
            }
          }
        }
      }
    }
  }
}